Issue #154 refactor: Ui Model 리팩토링
