+++ app/src/main/java/com/lateinit/rightweight/data/repository/RoutineRepository.kt
@@ -6,7 +6,7 @@ import com.lateinit.rightweight.data.database.entity.ExerciseSet
 import com.lateinit.rightweight.data.database.entity.Routine
 import com.lateinit.rightweight.data.database.intermediate.DayWithExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
-import com.lateinit.rightweight.ui.model.RoutineUiModel
+import com.lateinit.rightweight.ui.model.routine.RoutineUiModel
 import kotlinx.coroutines.flow.Flow
 
 interface RoutineRepository {
