+++ app/src/main/java/com/lateinit/rightweight/data/repository/UserRepository.kt
@@ -3,7 +3,7 @@ package com.lateinit.rightweight.data.repository
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
 import com.lateinit.rightweight.data.model.local.User
 import com.lateinit.rightweight.data.model.remote.WriteModelData
-import com.lateinit.rightweight.ui.model.HistoryUiModel
+import com.lateinit.rightweight.ui.model.history.HistoryUiModel
 import kotlinx.coroutines.flow.Flow
 import java.time.LocalDate
