+++ app/src/main/java/com/lateinit/rightweight/ui/model/ParentExerciseUiModel.kt
@@ -1,5 +1,7 @@
 package com.lateinit.rightweight.ui.model
 
+import com.lateinit.rightweight.ui.model.routine.ExercisePartTypeUiModel
+
 abstract class ParentExerciseUiModel(
     open var title: String,
     open val part: ExercisePartTypeUiModel,
