+++ app/src/main/java/com/lateinit/rightweight/ui/model/routine/DayUiModel.kt
@@ -1,4 +1,6 @@
-package com.lateinit.rightweight.ui.model
+package com.lateinit.rightweight.ui.model.routine
+
+import com.lateinit.rightweight.ui.model.ParentDayUiModel
 
 data class DayUiModel(
     val dayId: String,
