+++ app/src/main/java/com/lateinit/rightweight/ui/model/routine/RoutineUiModel.kt
@@ -1,12 +1,13 @@
-package com.lateinit.rightweight.ui.model
+package com.lateinit.rightweight.ui.model.routine
 
+import com.lateinit.rightweight.ui.model.ParentRoutineUiModel
 import java.time.LocalDateTime
 
-data class SharedRoutineUiModel(
+data class RoutineUiModel(
     override val routineId: String,
     override val title: String,
     override val author: String,
     override val description: String,
     override val modifiedDate: LocalDateTime,
-    val sharedCount: String
+    val order: Int
 ): ParentRoutineUiModel(routineId, title, author, description, modifiedDate)
