+++ app/src/main/java/com/lateinit/rightweight/ui/routine/detail/DetailExerciseAdapter.kt
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.lateinit.rightweight.R
 import com.lateinit.rightweight.databinding.ItemExerciseWithSetsBinding
-import com.lateinit.rightweight.ui.model.ExerciseUiModel
+import com.lateinit.rightweight.ui.model.routine.ExerciseUiModel
 
 class DetailExerciseAdapter(
     private val onClickExercise: (Int) -> Unit
