+++ app/src/main/java/com/lateinit/rightweight/ui/share/SharedRoutineFragment.kt
@@ -13,11 +13,10 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.NavHostFragment.Companion.findNavController
-import androidx.navigation.fragment.findNavController
 import com.google.android.material.snackbar.Snackbar
 import com.lateinit.rightweight.R
 import com.lateinit.rightweight.databinding.FragmentSharedRoutineBinding
-import com.lateinit.rightweight.ui.model.SharedRoutineSortTypeUiModel
+import com.lateinit.rightweight.ui.model.shared.SharedRoutineSortTypeUiModel
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
