+++ app/src/main/java/com/lateinit/rightweight/ui/share/SharedRoutineViewModel.kt
@@ -6,8 +6,8 @@ import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import androidx.paging.map
 import com.lateinit.rightweight.data.repository.SharedRoutineRepository
-import com.lateinit.rightweight.ui.model.SharedRoutineSortTypeUiModel
-import com.lateinit.rightweight.ui.model.SharedRoutineUiModel
+import com.lateinit.rightweight.ui.model.shared.SharedRoutineSortTypeUiModel
+import com.lateinit.rightweight.ui.model.shared.SharedRoutineUiModel
 import com.lateinit.rightweight.util.toSharedRoutineSortType
 import com.lateinit.rightweight.util.toSharedRoutineUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
