+++ app/src/main/res/layout/item_history_set.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="historyExerciseSetUiModel"
-            type="com.lateinit.rightweight.ui.model.HistoryExerciseSetUiModel" />
+            type="com.lateinit.rightweight.ui.model.history.HistoryExerciseSetUiModel" />
 
     </data>
